Just an update on things
update schematic to and pcb. figured out a way to user int, ext and chager over three pin connector with only to wires to PCB and no extra diode needed. getting closer to the anwers on PWM pins to use with what fuction.
very little let to do on PCB, but things have to get checked before getting it done.